Update students page

+++ src/components/CRUDStudent/DeleteStudent/index.js
@@ -4,10 +4,12 @@ import { db } from "../../../firebase";
 
 function DeleteStudent({studentID}) {
     const handleDelete = async () => {
-        //console.log(studentID);
-        const deleteStudent = await deleteDoc(doc(db,'Student',studentID))
-            .then(() => alert("Delete student " + studentID + " successfully"))
-            .catch(err => alert("Failed to delete student"));
+        if (window.confirm("Are you sure to delete student " + studentID + "?") == true) {
+            const deleteStudent = await deleteDoc(doc(db,'Student',studentID))
+                .then(() => alert("Delete student " + studentID + " successfully"))
+                .catch(err => alert("Failed to delete student"));
+        }
+        
     }
